OTP-8484 Fix minor cross compilation issues
The use of mmap() was unnecessarily disabled when cross compiling.

The configure arguments --with-ssl, and --with-odbc refused to accept
libraries outside of $erl_xcomp_sysroot when cross compiling for no good
reason.

The configure argument --with-odbc didn't handle the value yes correct.

The configure arguments --with-odbc, and --without-odbc have also been
added to the configure help.

(Thanks to Steve Vinoski for reporting these issues)