odbc: Remove support for old MACs
Change configure to skip odbc for old MACs, the change in
PR-1227 is not backwards compatible with old MACs, and we do not
see a need to continue support for such old versions. However
it is still possible to make it work on such machines using
the --with-odbc configure option.